move unikernel service code to pkg/unikernel/service
